Remove `script.sh`
`script.sh` was running in `%post` and was "sneakily" creating files
that weren't tracked by the RPM. This was causing hidden failure when
installing smart-mon alongside golang-github-prometheus-node_exporter.

This adds the systemd service file to the RPM spec, and removes the
`script.sh` file so everything is properly tracked in the RPM for Zypper
to understand.